fix(zk): proof compatiblity between 32/64b platforms

+++ tfhe-zk-pok/src/proofs/range.rs
@@ -225,7 +225,7 @@ pub fn prove<G: Curve>(
             core::slice::from_mut(s),
             &[
                 hash_s,
-                &i.to_le_bytes(),
+                &(i as u64).to_le_bytes(),
                 v_hat.to_le_bytes().as_ref(),
                 c_hat.to_le_bytes().as_ref(),
                 c_y.to_le_bytes().as_ref(),
@@ -336,7 +336,7 @@ pub fn verify<G: Curve>(
             core::slice::from_mut(s),
             &[
                 hash_s,
-                &i.to_le_bytes(),
+                &(i as u64).to_le_bytes(),
                 v_hat.to_le_bytes().as_ref(),
                 c_hat.to_le_bytes().as_ref(),
                 c_y.to_le_bytes().as_ref(),
